Move auxiliary actions to dedicated file
