+++ src/module/actor/character/data.ts
@@ -34,8 +34,8 @@ import { BaseWeaponType, WeaponCategory, WeaponGroup } from "@item/weapon/types.
 import { ValueAndMax, ZeroToFour } from "@module/data.ts";
 import { DamageType } from "@system/damage/types.ts";
 import type { Predicate } from "@system/predication.ts";
+import type { WeaponAuxiliaryAction } from "./auxiliary.ts";
 import type { CharacterPF2e } from "./document.ts";
-import type { WeaponAuxiliaryAction } from "./helpers.ts";
 import type { CharacterSheetTabVisibility } from "./sheet.ts";
 
 type CharacterSource = BaseCreatureSource<"character", CharacterSystemSource> & {
